- lcd.config: no yaw.D for pid.controller==1 + add amperage to MSP_ANALOG (as suggested by haydent, adapted for dev version layout) - lcd.telemetry: display remaining time, if armedtimewarning is set - lcd.telemetry: display remaining mAh, if capacity alarm treshold is set
git-svn-id: http://multiwii.googlecode.com/svn/trunk/MultiWii_shared@1542 02679b44-d973-9852-f2fa-63770883b36c

+++ MultiWii.h
@@ -108,7 +108,7 @@ extern int16_t lookupThrottleRC[11];
   extern uint16_t cycleTimeMax;       // highest ever cycle timen
   extern uint16_t cycleTimeMin;       // lowest ever cycle timen
   extern int32_t  BAROaltMax;         // maximum value
-  extern uint8_t  GPS_speedMax;       // maximum speed from gps
+  extern uint16_t GPS_speedMax;       // maximum speed from gps
   extern uint16_t powerValueMaxMAH;
 #endif
 #if defined(LOG_VALUES) || defined(LCD_TELEMETRY) || defined(ARMEDTIMEWARNING) || defined(LOG_PERMANENT)
